Shift to Osmium-based merging
rev-geocode-merge.py contains the new process for merging NBI and OSM data. It produces a separate OSM file with just the new NBI tags. A new OSM handler class was created to handle this function.

In the future, the created file will be a full merge of the two datas.